Add basic histogram for query execution
This commit adds a basic histogram for query execution times.
Currently the number of buckets is fixed to 10, and arranged in
the following way,
- each query has two fields, buckets and frequencies
- buckets is an array of upper bounds of each bucket (this is serial
histogram, hence only require upper bounds)
- frequencies array contains the count for each of the buckets
repsectively

With the current implementation, the first 10 distinct values are taken
as the upper bounds of the buckets. An extra condition is in place to
ensure that each of these boundaries maintain a difference of atleast 10%.

TODO:
- implement a dynamic repartition algorithm on these histograms to
make them equi-depth.